organized the folder structure of code
