+++ Pokemon/Pokemon.vcxproj
@@ -154,14 +154,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="BattleManager.cpp" />
-    <ClCompile Include="Game.cpp" />
-    <ClCompile Include="main.cpp" />
-    <ClCompile Include="Player.cpp" />
-    <ClCompile Include="Pokemon.cpp" />
-    <ClCompile Include="ProfessorOak.cpp" />
-    <ClCompile Include="Utility.cpp" />
-    <ClCompile Include="WildEncounterManager.cpp" />
+    <ClCompile Include="src\Battle\BattleManager.cpp" />
+    <ClCompile Include="src\Main\Game.cpp" />
+    <ClCompile Include="src\main.cpp" />
+    <ClCompile Include="src\Character\Player\Player.cpp" />
+    <ClCompile Include="src\Pokemon\Pokemon.cpp" />
+    <ClCompile Include="src\Character\ProfessorOak.cpp" />
+    <ClCompile Include="src\Utility\Utility.cpp" />
+    <ClCompile Include="src\Battle\WildEncounterManager.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Content Include=".idea\.idea.Pokemon\.idea\indexLayout.xml" />
@@ -170,16 +170,17 @@
     <Content Include="Pokemon.vcxproj.filters" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="BattleManager.h" />
-    <ClInclude Include="Game.h" />
-    <ClInclude Include="Grass.h" />
-    <ClInclude Include="Player.hpp" />
-    <ClInclude Include="Pokemon.hpp" />
-    <ClInclude Include="PokemonChoice.hpp" />
-    <ClInclude Include="PokemonType.hpp" />
-    <ClInclude Include="ProfessorOak.h" />
-    <ClInclude Include="Utility.h" />
-    <ClInclude Include="WildEncounterManager.h" />
+    <ClInclude Include="include\Battle\BattleManager.h" />
+    <ClInclude Include="include\Battle\BattleState.h" />
+    <ClInclude Include="include\Main\Game.h" />
+    <ClInclude Include="include\Pokemon\Grass.h" />
+    <ClInclude Include="include\Character\Player\Player.hpp" />
+    <ClInclude Include="include\Pokemon\Pokemon.hpp" />
+    <ClInclude Include="include\Pokemon\PokemonChoice.hpp" />
+    <ClInclude Include="include\Pokemon\PokemonType.hpp" />
+    <ClInclude Include="include\Character\ProfessorOak.h" />
+    <ClInclude Include="include\Utility\Utility.h" />
+    <ClInclude Include="include\Battle\WildEncounterManager.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
