+++ Pokemon/include/Battle/BattleState.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Pokemon.hpp"
+#include "../../include/Pokemon/Pokemon.hpp"
 
 struct BattleState
 {
