+++ Pokemon/include/Battle/WildEncounterManager.h
@@ -1,4 +1,4 @@
-#include "Grass.h"
+#include "../../include/Pokemon/Grass.h"
 #include <vector>
 
 class WildEncounterManager
