+++ Pokemon/include/Character/Player/Player.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <string>
-#include "Pokemon.hpp"
+#include "../../../include/Pokemon/Pokemon.hpp"
 
 class Player
 {
