+++ Pokemon/include/Main/Game.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Grass.h"
+#include "../../include/Pokemon/Grass.h"
 
 class Player;
 class Game
