+++ Pokemon/src/Pokemon/Pokemon.cpp
@@ -1,6 +1,6 @@
 #pragma once
-#include "PokemonType.hpp"
-#include "Pokemon.hpp"
+#include "../../include/Pokemon/PokemonType.hpp"
+#include "../../include/Pokemon/Pokemon.hpp"
 
 Pokemon::Pokemon()
 {
